refactor samples estructure
