--- helios-sample/gradle.properties
@@ -1,4 +1,0 @@
-# Maven publishing configuration
-POM_NAME=Helios Sample
-POM_ARTIFACT_ID=helios-sample
-POM_PACKAGING=jar
