+++ helios-samples/core-sample/gradle.properties
@@ -1,0 +1,4 @@
+# Maven publishing configuration
+POM_NAME=Helios Core Sample
+POM_ARTIFACT_ID=helios-core-sample
+POM_PACKAGING=jar
