+++ helios-samples/retrofit-sample/build.gradle
@@ -1,14 +1,13 @@
 apply plugin: 'kotlin-kapt'
 
-def arrow_version = "0.9.0"
-
 dependencies {
     compile project(":helios-core")
     compile project(":helios-optics")
     compile project(":helios-integrations:helios-retrofit")
-    compile "com.squareup.retrofit2:retrofit:2.6.0"
+    
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
-
+    compile "com.squareup.retrofit2:retrofit:$retrofitVersion"
+    
     kapt project(':helios-meta')
     kapt project(':helios-dsl-meta')
 }
