+++ helios-samples/retrofit-sample/gradle.properties
@@ -1,0 +1,4 @@
+# Maven publishing configuration
+POM_NAME=Helios Retrofit Sample
+POM_ARTIFACT_ID=helios-retrofit-sample
+POM_PACKAGING=jar
