+++ helios-samples/retrofit-sample/src/main/kotlin/helios/sample/retrofit/HttpBinApi.kt
@@ -1,0 +1,9 @@
+package helios.sample.retrofit
+
+import retrofit2.Call
+import retrofit2.http.GET
+
+interface HttpBinApi {
+  @GET("json")
+  fun getJson(): Call<Response>
+}
