+++ settings.gradle
@@ -6,8 +6,8 @@ include ':helios-meta'
 include ':helios-core'
 include ':helios-parser'
 include ':helios-benchmarks'
-include ':helios-sample'
-include ':helios-sample-retrofit'
+include ':helios-samples:core-sample'
+include ':helios-samples:retrofit-sample'
 include ':helios-optics'
 include ':helios-dsl-meta'
 include 'helios-test'
